fix: resolve exponential parse time on unclosed components
Previously "begin", all lowercase, was interpreted as a valid property key, but only when parsing as a component failed. This meant that in the failure case, where there were many unclosed and lowercase `begin`s, the parse time would increase at an O(2^N) rate, taking ~8s for 20 unclosed `begin`s.

With the new and improved `property_key()`, it should be case-insensitive and work for arbitrary properties that begin with `begin` or `end`, which would not before.