fix: replace HashMap with BTreeMap for properties, for determinism
I noticed that when serializing a calendar for my dependent project,
the order of properties was randomized on every run.

This is just like https://github.com/hoodie/icalendar/issues/7,
iterating over a HashMap is a common source of non-determinism, so
I just replaced it with a BTreeMap which always sorts its pairs.